CreateOSDCluster: Reset state before redirect
Greg noticed an issue that if you create a cluster, go back
to the cluster list, and then try to create a cluster again, instead
of getting the form you'd just be redirected immediately to Cluster
Details of the cluster you just created.

This is probably a race condition because this.reset() doesn't complete
before the <Redirect> is rendered. Calling reset() on
componentWillUnmount() fixes this problem by ensuring the state
is reset when we leave the page.