ClusterActionsDropdown: "Launch Console" should be an <a> element
By default, PF's `<DropdownItem>` renders an `<a>` element. However,
since most of the items in the dropdown cause actions within the app
(with `onclick` handlers), we set them all to use a `<button>` element
instead, as is best practice.

For consistency, each item got this prop (`component="button"`) from
a shared `baseProps` dictionary, and I didn't notice that "Launch
Console" is special and shouldn't inherit this prop.

Being a `<button>` and not an `<a>` meant that the `href` attribute
didn't work (obviously) and this prevented people from launching
the console via the dropdown in the cluster list.

This commit fixes this issue.

Fixes: https://jira.coreos.com/browse/SDA-1279