Limit display names to 63 characters, and support long display names
This MR limits the display name length to 63 characters. This is done
because supporting an arbitrarily long display name is a maintenance
burden, and a source for a lot of visual bugs in our apps.

In addition to limiting the display name, some CSS fixes were added to
ensure we properly support long cluster names / display names:
enabling word-break on the "cluster name" tooltip in the cluster list,
and the cluster name title in cluster details.

Also, the length of the cluster name column was set to a static 30% to
try to match the mockup in https://jira.coreos.com/browse/SDE-159

(unfortunately, PF4 does not allow us to specify table column widths in
grid column units like PF3 allowed us, so we have to hardcode a
percentage)

This should fix:
- https://jira.coreos.com/browse/SDA-1137
- https://jira.coreos.com/browse/SDA-1140
- https://jira.coreos.com/browse/SDA-1260