users: use URI encoding when deleting
If someone created a cluster user with a % character,
they won't be able to delete it because it was not escaped properly.

Using `encodeURIComponent` ensures the URI for deletion is properly
encoded, and the single % is escaped to %25, allowing users with
this character to be deleted.

This is an edge case, since most identity providers probably won't
allow users with this character, but fixing it doesn't hurt.

This fixes https://jira.coreos.com/browse/SDA-1004