Fix machine types sorting bug on Chrome, simplify code
The sorting function worked in Firefox (and on the tests!) by
coincidence only, and didn't work in Chrome at all.

Reading its code again, it was needlessly complicated, and didn't catch
cases where memory was equal, but CPUs weren't (and then you want the
machine types with less CPUs to be before the one with more CPUs)

Also, sorting on render() caused the sort function to be called
for every state transition, which is inefficient, and worse - caused
some race conditions with componentDidUpdate() selecting the default
machine type. Now we sort just once, when we get the response, before
we set the default.

Fixes https://jira.coreos.com/browse/SDA-1298