package.json: Restore lodash-es transformIgnorePattern
So it turns out that the way this regex works is "modules that match
this are NOT transformed". We want to transform lodash-es if it exists,
but we don't want to transform anything else. With the .*, it matched
ALL the modules, so they were all transformed, drastically slowing down
the tests.

The change with the .* was done because patternfly-table used lodash-es
in a weird way that broke the original regex, but now patternfly-table
changed to use regular lodash, so we can restore the previous regex.

If any dependency adds lodash-es in a weird way (not hoisted to the main
node_modules directory) we can rethink the regex.

In other words, "cross that bridge when we get to it", and until then
don't slow the tests too much :)