Add a yarnrc pointing to the internal Red Hat yarn registry
This is so our jenkins builds don't overload npmjs for requests.

PLEASE NOTE:
This will mean that you will not be able to run `yarn install` without
being inside the Red Hat network.

It also means that any machine that is building this code MUST have the
Red Hat internal IT certificate installed in the computer's trust
store.

If we don't want to set this in the repo directly, we'll have to ask
SREs to set it on Jenkins. For now, it seems easier to just set it in
the repo