Support filtering by a subscription's entitlement_status
This commit adds support on filtering by a subscription
entitlement_status, and setting this filter by a query parameter.

To filter by entitlement_status, we need to first call the Accounts
Management Service, get a list of subscriptions matching the relevant
criteria, collect the cluster IDs from these subscriptions, and then ask
Clusters Service for clusters with those IDs.

This is the reverse of the usual flow, where we ask Clusters Service for
clusters, and then collect list of subscription IDs for querying the
account management service.

Parsing the filter from the query is done in ClusterList, when it's
mounted, and the call to `.refresh()` on mount is delayed until
the redux state updates with the filter flags.

https://jira.coreos.com/browse/SDA-1294