config: Check that config environment exists before calling promise
When overridding the /beta environment with the `?env=` query string,
the local storage gets set in a way that the production environment will
attempt to read it, thus causing the website to break, since only the
production environment is available. We can mitigate this by just
verifying that the environment requested by the localStorage is
available in the configs map.