Show `Subscribed` as subscrption status for OSD
Currently the accounts management service returns always a blank
`entitlement_status` attribute for OSD subscriptions. As a result we
display _Unknown_ in the grid. But we should be displaying _Subscribed_.
This patch changes the `SubscriptionStatusIndicator` component so that
it receives the cluster as a property (instead of just the subscription)
and so that it renders _Subscribed_ instead of _Unknown_ when the
`entitlement_status` is missing or empty and the cluster is managed.

Related: https://jira.coreos.com/browse/SDA-1355
Signed-off-by: Juan Hernandez <jhernand@redhat.com>