CreateOSDCluster: Show spinner until quota, regions, and types are ready
Instead of showing the form with two spinners waiting for cloud
providers and machine types (and quota), show just one spinner,
and then show the form only when this data is fully loaded.

This looks better, matches the UX guidelines more closely, and prevents
the user from being able to click Create before they even see the
option to select a region and a machine type.

Also, while I'm at it, fixed the selector for the "flat radio buttons"
used in machine type selections, so the text will be properly aligned to
the left.

Fixes https://jira.coreos.com/browse/SDA-1328