Support `external_id` as cluster identifier in link to details page
Support `external_id` as cluster identifier in link to details page by
detecting for the ID to be an external ID / UUID, and if so, fetching
the cluster by that instead. Once we have the cluster by external ID,
react-router redirect (aka simple route change) to the regular ID.

The patch adds the uuid-validate module, and the RH IT CA for connecting
to Nexus to retrieve the dependency.