Search/sort/page via Accounts Management Service, not clusters service
Previously we queried cluster service first, collected list of
subscription IDs, and then searched for these IDs in AMS to grab
subscription info for the clusters.

This approach started breaking when we wanted to filter by fields
that are stored in the Accounts Management Service, so now we're moving
display_name to Accounts Management Service, and doing
searching/sorting/paging in Accounts Management Service.

We now collect a list of cluster IDs from the subscription objects,
and then query the cluster service for clusters with these IDs.

https://jira.coreos.com/browse/SDA-1382