Monitoring: Don't render a spinner when request is pending
We already show a spinner at the top bar of ClusterDetails,
and UI design guidelines call for just one spinner per page.

Also, rendering a spinner instead of the content causes the view to
reset and sections to collapse.

We can just trust the spinner at the top, and not monitor the pending
state inside the monitoring tab.

Changed the reducer to keep old data in state, and added an unmount handler
to clear the state like we did for logs, so we won't show irrelevant
data when moving to a different clusters.

This should fix https://jira.coreos.com/browse/SDA-1422