Remove `cluster.creator`
This attribute is about to be removed by the backend. It isn't in use in
the UI currently, but it is still in some of the test files. This patch
removes it completely from the UI.

Related: https://jira.coreos.com/browse/SDA-1395
Related: https://gitlab.cee.redhat.com/service/uhc-clusters-service/merge_requests/1201
Signed-off-by: Juan Hernandez <jhernand@redhat.com>