clusterActions: generate a 404 error when cluster is not found by UUID
When navigating to a cluster by UUID, the search does not return 404
when nothing is found, but rather an empty list. This confused the code
in `fetchSingleClusterAndPermissions` and other places, which expects
a cluster to not be undefined.

Generating a 404 error in this case allows the special 404 handling in
cluster details to work for the "navigate by UUID" usecase.

https://jira.coreos.com/browse/SDA-1428