RegisterCluster: move custom validators out of render()
While it seems like it makes sense to define custom validators in
render(), it can sometimes cause problems on user input.

User input triggers an update, which triggers render(), and then
the functions get re-defined (with a different location in memory, so
they're not equal in react's eye). This causes react to think the
validator function changed, so it runs the function and triggers another
update, which can trigger another update, and so on, until React bails
out with an invariant error.

This should fix https://jira.coreos.com/browse/SDA-1469