Add the ability to filter OCP / OSD clusters on the Cluster List
Like the subscription entitlement_status filter this too can be accessed
via a URL parameter, for example `?plan_id=OCP` or `?plan_id=OCP,OSD`.

This MR is also a small refactor of the filter dropdown and filter chipgroup,
and of `createViewQueryObject()`, to make it easier to add more filters
like these in the future.

Also, added tests for `createViewQueryObject()` and changed the Type
column values to be "OCP" and "OSD" for consistency with the filter
dropdown.

Fixes https://jira.coreos.com/browse/SDA-1474