Cluster List Filter Dropdown: Update query params when filter is changed
Update the query params displayed in the URL bar, for consistency. This
makes things less confusing if you got to the cluster list via the
subscriptions page, and then removed the filter. Before this change, the
query params would remain in the URL bar even when no filter was
applied.

Additionally, fix the loading state logic for cluster list so it won't
switch to the initial loading state if a filter is changed after the
previous filter returned an empty result. This should prevent the
dropdown from disappearing in these cases and improve the user
experience.

The loading state logic still needs to be improved because it'll still
switch to a loading state if you get an empty result and then remove ALL
filters.

https://jira.coreos.com/browse/SDA-1540