package.json: don't mark CSS as side effects
This seems to solve the problem that caused the CSS order to be wrong,
causing PatternFly styles to override our own styles.

Also, remove the include of Content in main.scss, since that style is
already included by PatternFly itself and including it there causes it
to be in the result twice.

This should fix https://jira.coreos.com/browse/SDA-1386