ArchivedClusterList: invalidate on mount
The usual flow for getting into archived clusters calls
invalidateClusters() when clicking on the menu item, but instead of
relying on the page that links to /archived to call this function, it
should happen on mount, to make sure it always happens.

Otherwise, the list in memory is considered "valid" even if it was valid
for active clusters and not archived ones, so the archived list won't be
fetched again.

This should fix https://jira.coreos.com/browse/SDA-1557