push_to_insights: Also push to the new qaprodauth staging environment
The new staging environment is https://qaprodauth.cloud.redhat.com/

This MR begins the process of aligning ourselves with the Insights
Platform regarding the use of the /beta environment.

As a first step, we'll push staging both to /beta and to qaprodauth.
We'll then wait a while to make sure it's all working properly and that
QE are comfortable with the new environment.

Afterwards, we'll change /beta to also point to our stable branch, just
like prod.