new staging env: Fix /beta path handling
On the new stage, we're going to have both `/openshift` and
`/beta/openshift` available to us - the first running with production
Insights Platform, and the second running with the Beta insights
platform.

This means we need to support both `/openshift` and `/beta/openshift` as
our basename in staging, otherwise `/openshift` won't work on the new
staging env.

Additionally, we'll need to support `/beta` in production later (once
the move to the new staging env is complete), so which API server our
app is using should be independent of the basename setting.

This MR introduces a new "--staging=true" flag for our build script,
an APP_STAGING global variable to go with it, and changes the push
script to build twice for staging - once with `/beta`, once without.