AddOns: Determine when to show Add-ons tab based on quota
Revamp quota calculations to include add-ons. If an add-on does not
exist in the quota summary, it should not be visible for that org's
clusters. If the add-on is in the quota summary, then we determine
whether to show it as enabled or disabled within the tab.

If no add-ons are available in the quota summary, the Add-ons tab should
be completely hidden from that org.