Cluster List: Use the insights skeleton loading pattern
When we don't have data yet, instead of showing just a spinner,
we can show all the page elements, including the table, with the
insights platform `<Skeleton>` component as placeholder for the actual
data.

This has two main benefits: it allows users to click Create before the
list is loaded, saving a second or two, and it makes loading look less
"jittery" since less elements change on the page and the layout stays
roughly the same.

Fixes https://issues.redhat.com/browse/SDA-1344

I also intend to do the same for Cluster Details later.