Change form dropdowns to use PatternFly's FormSelect
the `<FormSelect>` component uses the browser's native `<select>`
implementation. This means it requires a bit less boilerplate code, and
behaves better inside modals. I was advised on the PF slack that unless
we need any of the advanced features of the `<Select>` component, we
should use `<FormSelect>`.

This fixes https://issues.redhat.com/browse/SDA-1774