Update eslint and linter rules to the latest version
Most notable change is enforcing use of the shorthand `<>`
instead of the long form `<React.Fragment>`, and better propTypes validation.

I've modified the configuration slightly:
- links with `target="_blank"` without `rel="noopener noreferrer" now
produce a warning. I'll update them in a followup and then change the
configuration to error.

- deviating a bit from the airbnb style: while airbnb now prefers that
state is only initialized in a constructor, the preference in our team
is to avoid boilerplate. Marked state initialization in a constructor as
a warning, and in a followup later will change it to error.

- The new airbnb style enables the "no prop spreading" rule, however
prop spreading is useful and we use it a lot, so I disabled this rule

- The new airbnb style changed the `allow-parens` setting to "always",
meaning arrow functions will always need parens around arguments. Since
this change would've caused this already big commit to be even bigger,
I've added an override to keep it the way we had it `as-needed`, meaning
only eforcing parens where they're really needed. This can be changed in
a followup if we decide to have parens everywhere.