Reuse list data for quicker cluster details loading
The backend returns all (or at least almost all) details of clusters
when we ask for the cluster list. This means that we can reuse this data
to render the Cluster Details screen faster, minimizing latency!

When the user clicks on a cluster name on the list, we can set its data
in redux before navigating, then upon navigation the cluster details
page will already have all info rendered!

The cluster details page will still try to fetch the cluster details
when the page is loaded this way, for future-proofing in case the
backend will decide to return less data in the list itself. Loading
cluster details via direct link will still behave just like before.