CreateOSDCluster: implement node count input
Instead of a numeric input, node count is now a `<FormSelect>`,
and only allows the user to select amounts they have quota for.

If the user has no quota for extra nodes, the component is disabled and
the value is set to the base cluster size.

For Multi AZ, the component allows the user to select the number of
compute nodes per zone, and presents the calculation of how many total
nodes the cluster will have.

Additionally, this MR removes unnecessary calls to
`getOrganizationAndQuota()` that caused multiple requests to
`quota_summary` upon navigation to the create cluster page / scale
cluster dialog, and renamed `nodeQuota` in the redux state to
`clusterQuota`, to better reflect the meaning of the values there.

Should fix: https://issues.redhat.com/browse/SDA-1731