Scale cluster: disable "apply" until fields are touched
Using the redux-form `pristine` prop we can easily disable the "apply"
button if the fields were not touched by the user.

Also, using `enableReinitialize: true` in the redux form config allows
us to drop some confusing code that basically re-implemented this
functionality. This is not only a refactor for code clarity - it is
required for `pristine` to work in the case of this form (without the
refactor, calling `change` would've changed `pristine` to `false`.

Fixes: https://issues.redhat.com/browse/SDA-1864