Validators: remove issuer URL regex
https://issues.redhat.com/browse/SDA-1885

The JIRA ticket asked to simplify the regex. What's a better way of
simplifying it than to remove it completely? Now we just check the URL
is valid, has no hash/query, and starts with https.

Some of the test cases removed were of technically correct URLs
(apparently you can configure a DNS server to resolve empty names, so
`https://...` and the like are technically valid) that will probably
never work, but I doubt anyone will make this kind of mistake,
especially when the issuer URL is usually copy-pasted and not written by
hand.