UnarchiveCluster: set new status to Disconnected
A cluster's status before it is archived is not retained.  Until
telemetry is received, we cannot know with certainty whether a newly
unarchived cluster should be Active or Disconnected.

If the status is set incorrectly, it will transition to the correct
state eventually.  Both Active and Disconnected work.  However, an
incorrect Disconnected status will reconcile much more quickly than an
incorrect Active status.  Thus, defaulting to Disconnected is preferred.

This resolves SDB-874.