Disable 'cancel' button when form is pristine in networking tab
This is to prevent users from resetting the form to the old initial values that have not yet
been updated in the store to the new values post update. In the meantime the initial values are updated
according to the submitted form, so the resetted form is marked dirty, creating an inconsistency.

After pressing 'save changes' in the modal, the form is set to be pristine and both buttons are disabled.

https://issues.redhat.com/browse/SDA-2214