TabsRow: Work around PatternFly bug causing empty tab content
PF bug:
https://github.com/patternfly/patternfly-react/issues/4095

Basically, we use the "separate content" feature of this react component
to achieve our visual style in Cluster Details. The problem is that PF
probably only tested this mode under the assumption that the available
tabs are static and set on mount. In our case, some tabs may appear
after the initial mount, and a bug on PF side causes the tab button
itself "steal" the ref we wanted to use for the tab contents, and claim
these refs for itself.

The result is, that when the tab was clicked, the component tried to
change the `hidden` property on the tab content ref, but instead changed
it on the button only.

To work around this, I had to implement the tab switching myself instead
of relying on PatternFly to do it, so the actual content ref is never
passed to the button, and thus never claimed by the wrong component.

https://issues.redhat.com/browse/SDA-2223