Networking tab: synchronous saving
The backend has a data consistency problem when many request to modify
privacy / ingresses happen in quick succession. This MR addresses this
by changing the requests to be synchronous.

async-await is a very useful tool to do this kind of thing, so this
turned out simpler than I expected.

I also changed it to dispatch a change to the cluster resource once the
request to modify it completes, which helps avoid the flicker QE
complained about in https://issues.redhat.com/browse/SDA-2216