installLinks: fix for IBM-Z
Same issue that we had with RHV. The consts were out of sync with the
channel dictionary that they were supposed to describe.

Moved the consts closer to the dictionary so it'll be easier to spot
errors like this, and changed the dictionary to
use the consts directly instead of hardcoding the keys.