Move to fetching unhealthy clusters (for 'Clusters with Issues' section) from '/dashboards/summary/unhealthy_clusters'
To support fetching "Unhealthy Clusters" (i.e. clusters with `health_state='unhealthy'`) for the `Clusters with Issues` section
we have created a designated endpoint in CS: `/dashboards/summary/unhealthy_clusters`.

This PR moves the query for those clusters to use the designated endpoint and, in addition, it moved the corresponding result set in the
state tree (`dashboardClusters`) from under `clusters` to under `dashboards` - this allows us to utilize `immer` which is used in the `dashboardsReducer`.

Signed-off-by: nshneor <nshneor@redhat.com>