RHMI: Enable RHMI product creation form
Since RHMI is a layered product on top of OSD, we can reuse most of the
OSD components for cluster creation. The few differences are accounted
for by hiding certain form fields that customers should not have access
to, as well as bypassing the cloud provider selector screen, since only
AWS is supported at the moment.