Add basic automatic tests for all pages
This MR adds basic snapshot testing to all pages. This is not a
replacement to component level tests, and does not test all possible
states of each page, but it still increases our coverage.

This testing is done with deep rendering of the entire component tree,
including redux. This is useful because we have a lot of logic in
untested redux `index.js` files, and because shallow rendering does not
catch errors in props passed to library components.

This test passing for a page doesn't mean the page works - it means that
the initial render of the page works. Pages can still break if they, for
example, handle network request state changes incorrectly (or any other
reason). This means these tests can't be relied on to know everything is okay,
but they can help us catch basic errors that are easy to miss.