Cluster List filter dropdown: disable when pending
Should fix https://issues.redhat.com/browse/SDA-2117

The issue seems to be a race between the filter and the refresh button.
If a refresh started and hasn't finished yet, it'll race the filtered
request. It seems to only be an issue with really slow connections.

Ideally, I'd prefer to keep it enabled and have it cancel the in-flight
request when the viewOptions change, but redux-promise-middleware
doesn't allow that, and I don't want to do a major rewrite just for
this.

Disabling the filter when a refresh is in progress should prevent this
issue from occurring, or at least should make it much less likely.