Backend proxy: support per-user tokens
This MR extends the backend proxy to support different tokens per user.
You can still use it like you always did, with one token for everyone,
but if you use the new `tokens_map` config file section, you can map
usernames to tokens, so the token used will depend on the user you
logged in with to the QA auth server.

If a token is specified in token_map for the current user,
it takes priority over the other two options (global token in env var or
global token in config file).

This MR also adds support for specifying a config file using the
enviornment variable `BACKEND_CONFIG`, so you can use different config
files with the new all-in-one `yarn start`

Fixes
https://issues.redhat.com/browse/SDA-2008
https://issues.redhat.com/browse/SDA-2330