[SDA-1969] Show logs when installing a cluster
Cluster installation logs will be shown in cluster details (where the
donuts usually go) when the cluster is installing. The log is refreshed
faster than the rest of the view, as the streaming log acts like
progress indication, to make it clear to the user installation isn't
stuck, it's just doing a lot of work.

https://issues.redhat.com/browse/SDA-1969