[SDA-1582] Integrate with Sentry
Integrate with our Sentry service for error reporting.
At the moment this includes catching any unhandled exception or anything
that leads to ErrorBoundry. More places to catch (like when there's an
error fetching data and such) can be added later.

Using a custom redux middleware, we also collect the history of
redux events leading up to the error, and recording useful metadata
of some redux events like the current location or which modal was opened.

https://issues.redhat.com/browse/SDA-1582