Check cluster status while installing, show error if errored
This MR adds a new component, the ClusterStatusMonitor. This component
called the lightweight `/clusters/<id>/status` endpoint every 5 seconds
when the cluster is pending or installing.

If the status changes, it asks the cluster details view to refresh (so
it will refresh when the cluster finishes installing),
and in case the status is "error" a PatternFly `<Alert>` will be shown
with the error description from the `/status` endpoint.

https://issues.redhat.com/browse/SDA-2438