RHMI: Enable CCS clusters
The current state of RHMI is that SREs create CCS OSD clusters on behalf
of the customer and install the RHMI add-on. Since the customer cannot
yet do self-service RHMI installations, we enable CCS so that SREs can
create RHMI clusters directly, rather than RHMI-on-OSD. Once quota
reconciliation is fully in place, users will then be able to create CCS
RHMI clusters.