Use babel-plugin-transform-imports for better treeshaking
Following the recommendations from insights-frontend-components, we can
achieve significant bundle size reduction by using
`babel-plugin-transform-imports` to import cjs versions of some of our
dependencies.

Before: ~7.65 MiB
After: ~4.27MiB

The usual way in which we import a component from PatternFly or
insights-frontend-components it imports the entire library into our
codebase, and while webpack's terser is not a bad tree-shaker, it's just
not good enough to properly remove all this dead code.

But with `babel-plugin-transform-imports` and the mappers in
babel.config.js, we can import the specific modules we use - and nothing
else!

This change does mean that in some cases, when you want to import
some helper functions/consts from PatternFly, you might need to add a
mapper for them in `babel.config.js`. This is because these
helpers/consts sometimes live in files together with a parent component,
so the name with import is not the name of the file. See the comment in
`babel.config.js` for more details.

*`treeshaking`, - dead-code elimination, but in the JavaScript world