[SDA-2500] Uninstallation logs
Add support for showing uninstallation logs while the cluster is
uninstalling. When a user deletes a cluster, they still get redirected
to the cluster list as before, but if they open the details page while
the uninstalltion is in progress, they'll see the uninstallation log.

When the cluster is done uninstalling, the user will be redirected back
to the cluster list automatically, and a notification will be shown
saying the cluster was uninstalled successfully.

In addition to supporting uninstallation logs, I also fixed the
fullscreen style bug on Firefox (SDA-2454),
added some defensiveness around the use of the logPaneRef in ScrollToEnd
to avoid a rare edge case spotted on Sentry where the ref would be
undefined, and added document visibility state handling so that the
log view will keep auto-scrolling if you switch to another tab and go
back, and also reduced the update frequency when the tab is not visible
for performance reasons.

https://issues.redhat.com/browse/SDA-2454
https://issues.redhat.com/browse/SDA-2500