[SDA-2572] Update react-redux, redux-form and misc dependencies
This is not just a regular update to keep up with our dependencies. It
also gives us the ability to write redux-powered components using React
Hooks, and slightly improves our app's performance.

As a side effect of new react-redux I had to remove snapshot testing for
the Router test, and make it only test that the component rendered with
no exceptions (without snapshot comparison).

This might not be ideal,
but react-redux causes weird issues with enzyme when snapshotting
connected components, and causes it to crash due to an out of memory
error. This error seems to be caused by a cyclic reference inside the
rendered component tree, with internal react-redux properties (that
aren't even exposed in the final snapshot) breaking the snapshot
formatter. I spent an entire day trying to fix it and made no progress.

My hope is that the boolean "rendered or not" without snapshot
comparison will be enough, and that we'll instead focus on having
a good enough test coverage with both jest unit tests per component,
and selenium integration tests.