Cluster details: ErrorBoundary inside each tab
Currently, when anything inside cluster details crashes, the whole
page is replaced with "Something went wrong" and user can only use
left sidebar to navigate back to clusters list (or Back button).
This will try to contain crashes to a single tab, rendering the
heading and allowing to access the other tabs.
(We may still get the previous whole-page error if the logic
computing _which_ tabs to show crashes.)