Deploy to ci branch to test stage SSO
There's an SSO upgrade coming up, and we want to test it. A new API
environment was created, protected by the VPN and an SSH tunnel, which
has our backend configured to authenticate with staging SSO instead of
production SSO.

We can't use the normal "qa" branch of Insights for this environment,
because we already use it (we use "qaprodauth" which uses the qa branch)
to talk to our staging backend, which uses production sso.

So let's try the CI branch. To support more than one alternative default
API environment, I changed the APP_STAGING variable to be APP_API_ENV,
and it now accepts a string that is the API environment identifier in
`config.js`.