Allow using ?env=stageSSO as environment override in prod
To test stage SSO, we can't use the Insights "ci" environment. That uses
"qa" SSO (which is not the same!). We have to use
https://cloud.stage.redhat.com

(the proxy file from https://mojo.redhat.com/docs/DOC-1213497 is
required on the user's machine to access that environment)

The cloud.stage.redhat.com environment mirrors prod-stable / prod-beta.
So, to properly test stage SSO, we need to include its config file in
production mode, which will allow using the `?env` environment override
feature to switch to the backend that uses stage SSO.

The plan is to merge this to master, verify staging still works (it
should not be affected), then pushing this to candidate and testing it
on https://cloud.stage.redhat.com/beta/openshift

Once testing is complete, we'll delete this configuration.