Track the Insights Platform deployment Jenkinsfile in our repo
This file used to be tracked in the deployment repo, making it
difficult to update.
https://github.com/RedHatInsights/uhc-portal-frontend-deploy/blob/prod-beta/58231b16fdee45a03a4ee3cf94a9f2c3

What we have there is a very old version of the file, and it
only exists in prod-stable/prod-beta and doesn't exist in the `qa`
branches.

Without this file in the qa branches, it's not possible to reliably push
to cloud.stage.redhat.com

Instead of excluding this file from being updated when we push to the
deployment repo, it's much better to track this file here, so we can
update it if needed and make sure it exists in all branches.