[SDA-2799] Validate AWS Account ID in byoc form
Turns out "AWS account ID" is misleading. In SDA-2799 you can see a bug
report from a user who assumed the account ID should be the string ID of
the account, and not the numeric ID. This caused confusion, because the
error from the backend just said `The credentials provided don’t belong
to Account ID aws-acm-demo`.

This commit adds an extended help text to the component with a
documentation link, as well as a validator for the 12 digits numeric ID.

I also created a new component, `ExternalLink`, to avoid needing to
repeat `target="_blank"` and `rel="noreferrer noopener"` everywhere we
have a link that leads out of OCM.

In a followup MR I'll change all
external links to use this component, so they will all have the
external link icon as Cameron requested in another ticket.

https://issues.redhat.com/browse/SDA-2799