[SDA-2833]: Fix overview crash once again
This is the same as 931694b77122b2037966e843ffca42992773cfdf which got
accidentally reverted in a bad rebase.

The isDialog prop must be removed, this kind of prop is way too error
prone.

In general, we should treat props and components as API - which means if
possible don't change the default behaviour of a component, and don't
add new required props unless you absolutely have to.