TitleWithLink: manual implementation for scrolling on load
We added anchors so try.openshift.com can link to specific sections in
the OCP instructions page. This didn't work for logged-out users - on
log in, they would get the top of the page even if there was an anchor
in the URL.

This manual implementation of scrolling on first render fixes this case
for users who are just logging in, manually scrolling to the relevant
title.