[SDA-2897] Installation log view performance fixes
Now that the log view is collapsible/expandable, it makes no sense to
poll logs in the background when the view is invisible.
That slows everything down with many requests, and many re-renders,
slowing everything down.

I moved the `ExpandableSection` into `InstallationLogsView` so the view
can know when it's expanded and when it's not, and turn off the timer
when it's invisible.

For an additional performance boost, I moved the responsibility of
counting the number of lines in the log from the component to the
reducer, so this number is only calculated once new data is fetched.
Before this change, the component was doing `.split('\n')` on the log
lines every time it rendered and every time it needed to fetch more
lines, and this can get very slow as the log grows bigger - especially
in the render path!

I also removed the big vertical gap above the view, which is not needed
now that it's expandable.