[SDA-2934] Monitoring: Don't show last check-in time when installing
In commit 9bcc0630fdede42ae98dfac98613bdb8b5dd97cd,
part of https://gitlab.cee.redhat.com/service/uhc-portal/-/merge_requests/1647
the logic for when the last check-in timestamp was changed to always
show it, but the prop `lastCheckIn` was not added in the `isInstalling`
case in `Monitoring.jsx`, causing it to be `null`, thus showing as if
the cluster last checked in at 1970-01-01.

This timestamp should be hidden for the installation case, because if
the cluster is still installing it probably haven't checked in yet, and
`activity_timestamp` will be the creation timestamp.

This should fix https://issues.redhat.com/browse/SDA-2934